Separate tests into separate files
